requirements.txt
Flask==3.0.0
gunicorn==21.2.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.0
requests==2.31.0
python-dotenv==1.0.0
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.33.3
Pillow>=10.0.0
